refactor: rename options class
